better fake data & run utests
